handle package released messages
