+++ investigator/investigator/package_release/__init__.py
@@ -1,0 +1,24 @@
+#!/usr/bin/env python3
+# thoth-investigator
+# Copyright(C) 2020 Bissenbay Dauletbayev
+#
+# This program is free software: you can redistribute it and / or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+"""The important parts for parsing package_release messages."""
+
+from .investigate_package_release import parse_package_release_message
+
+__all__ = [
+    "parse_package_release_message",
+]
